Semrush integration (#5)
* openai integration

* openai: remove unused model

* refactor controller package

* Semrush integration

* Semrush integration, more APIs

* adding keyword analytic API

* Adding unit tests

* update traffic API

* Updates after Semrush testing

---------

+++ src/main/java/com/searchify/suggestion/api/response/TrafficGeoDistResponse.java
@@ -20,7 +20,7 @@ import java.time.YearMonth;
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public class TrafficGeoDistResponse {
     private String countryCode;
-    private Integer traffic;
+    private Long traffic;
     private Double trafficShare;
     private Double desktopShare;
     private Double mobileShare;
