+++ src/main/java/com/searchify/suggestion/api/response/TrafficSummaryResponse.java
@@ -1,6 +1,5 @@
 package com.searchify.suggestion.api.response;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -8,8 +7,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-
-import java.time.YearMonth;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -20,16 +17,16 @@ import java.time.YearMonth;
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public class TrafficSummaryResponse {
     private String target;
-    private Integer visitAmount;
-    private Integer desktopVisits;
-    private Integer mobileVisits;
+    private Long visitAmount;
+    private Long desktopVisits;
+    private Long mobileVisits;
     private Double pagesPerVisit;
     private Double desktopPagesPerVisit;
     private Double mobilePagesPerVisit;
     private Double bounceRate;
     private Double desktopBbounceRate;
     private Double mobileBbounceRate;
-    private Integer userAmount;
+    private Long userAmount;
 
     @JsonFormat(pattern = "yyyy-MM")
     private YearMonth displayDate;
